backend/requirements.txt
fastapi
uvicorn
pymongo
python-multipart
pillow
colorgram.py
opencv-python
python-dotenv
rembg
numpy<2.0.0
scikit-learn
